requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite
pymongo
Pillow
requests
plotly
